+++ schema/postgres/v1.10.1.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "hit" ALTER COLUMN "screen_class" TYPE varchar(5);
+ALTER TABLE "screen_stats" ALTER COLUMN "class" TYPE varchar(5);